backend/requirements.txt
fastapi
uvicorn
rendercv[full]
google-generativeai
python-multipart
pyyaml
python-dotenv
